Init types refactor

+++ operators.py
@@ -21,23 +21,27 @@ import site
 from threading import Thread
 import time
 import webbrowser
+from bpy.app.handlers import persistent
 
 from .data import (
+    DSAccount,
     TrackingEvent,
     UIContext,
-    InitSource,
+    InitType,
     OutputDisplayLocation,
     RenderState,
     copy_image,
     format_rest_args,
+    get_anim_images,
     get_init_image_dimensions,
-    get_init_source,
+    get_init_type,
     get_preferences,
+    get_settings,
     initialize_sentry,
     log_sentry_event,
 )
 from .dependencies import install_dependencies, check_dependencies_installed
-from .requests import log_analytics_event, render_img2img, render_text2img
+from .requests import get_account_details, log_analytics_event, render_img2img, render_text2img
 import multiprocessing as mp
 import threading
 from glob import glob
@@ -57,17 +61,24 @@ def open_folder(dir: str):
 
 # Create and clear render directories. This function should get all filesystem info for rendering,
 # as well as any platform specific info.
-def setup_render_directories():
+def setup_render_directories(clear_anim=True):
     dreamstudio_dir = os.path.join(tempfile.gettempdir(), "dreamstudio")
     rendered_dir = os.path.join(dreamstudio_dir, "rendered")
     generated_images_dir = os.path.join(dreamstudio_dir, "generated_images")
     generated_animation_dir = os.path.join(dreamstudio_dir, "generated_animation")
-    for dir in [dreamstudio_dir, generated_images_dir, rendered_dir, generated_animation_dir]:
+    for dir in [
+        dreamstudio_dir,
+        generated_images_dir,
+        rendered_dir,
+        generated_animation_dir,
+    ]:
         if not os.path.exists(dir):
             os.mkdir(dir)
         if not os.access(dir, os.W_OK):
-            raise Exception(f"Directory {dir} is not writable. Please check your Blender application permissions.")
-        if dir == generated_animation_dir:
+            raise Exception(
+                f"Directory {dir} is not writable. Please check your Blender application permissions."
+            )
+        if dir == generated_animation_dir and clear_anim:
             for file in glob(os.path.join(dir, "*")):
                 os.remove(file)
     return rendered_dir, generated_images_dir, generated_animation_dir
@@ -124,7 +135,6 @@ class DS_SceneRenderExistingOutputOperator(Operator):
         DreamStateOperator.ui_context = UIContext.SCENE_VIEW
         DreamStateOperator.render_state = RenderState.RENDERING
         DreamStateOperator.render_start_time = time.time()
-        DreamStateOperator.rendering_from_texture = True
         bpy.ops.dreamstudio.dream_render_operator()
         return {"FINISHED"}
 
@@ -137,7 +147,7 @@ class GeneratorWorker(Thread):
         ui_context: UIContext,
         input_img_paths: List[str],
         output_img_directory: str,
-        init_source: InitSource,
+        init_type: InitType,
     ):
         self.scene = scene
         self.context = context
@@ -145,7 +155,7 @@ class GeneratorWorker(Thread):
         self.input_img_paths: List[str] = input_img_paths
         self.output_img_directory = output_img_directory
         self.running: bool = True
-        self.init_source: InitSource = init_source
+        self.init_type: InitType = init_type
         Thread.__init__(self)
 
     def run(self):
@@ -174,7 +184,7 @@ class GeneratorWorker(Thread):
         DreamStateOperator.last_rendered_image_path = output_file_path
 
         # text2img mode
-        if self.init_source == InitSource.TEXT:
+        if self.init_type == InitType.TEXT:
             DreamStateOperator.render_state = RenderState.DIFFUSING
             status, reason = render_text2img(self.output_img_directory, args)
             DreamStateOperator.render_state = RenderState.FINISHED
@@ -183,7 +193,7 @@ class GeneratorWorker(Thread):
             return
 
         # img2img mode - image editor, which can only generate from textures and text
-        if self.init_source == InitSource.EXISTING_IMAGE or self.init_source == InitSource.TEXTURE:
+        if self.init_type == InitType.TEXTURE or self.init_type == InitType.TEXTURE:
             DreamStateOperator.render_state = RenderState.DIFFUSING
             if not os.path.exists(init_img_path):
                 raise Exception(
@@ -191,16 +201,14 @@ class GeneratorWorker(Thread):
                         init_img_path
                     )
                 )
-            status, reason = render_img2img(
-                init_img_path, output_file_path, args
-            )
+            status, reason = render_img2img(init_img_path, output_file_path, args)
             if status != 200:
                 raise Exception("Error generating image: {} {}".format(status, reason))
             DreamStateOperator.render_state = RenderState.FINISHED
             return
 
         # img2img mode - 3D view
-        if self.init_source == InitSource.VIEWPORT:
+        if self.init_type == InitType.VIEWPORT:
             input_img_path = self.input_img_paths[0]
             if not os.path.exists(input_img_path):
                 raise Exception(
@@ -211,7 +219,7 @@ class GeneratorWorker(Thread):
             status, reason = render_img2img(input_img_path, output_file_path, args)
             if status != 200:
                 raise Exception("Error generating image: {} {}".format(status, reason))
-        elif self.init_source == InitSource.EXISTING_VIDEO:
+        elif self.init_type == InitType.ANIMATION:
             rendered_frame_image_paths = list(sorted(self.input_img_paths))
             if len(rendered_frame_image_paths) == 0:
                 raise Exception(
@@ -225,13 +233,18 @@ class GeneratorWorker(Thread):
                     or DreamStateOperator.render_state == RenderState.CANCELLED
                 ):
                     return
+                print('frame set', i + 1)
                 scene.frame_set(i + 1)
+                print('did set frame')
                 DreamStateOperator.render_start_time = time.time()
                 args = format_rest_args(settings, scene.prompt_list)
+                print('args', args)
                 output_file_path = os.path.join(
                     self.output_img_directory, f"result_{i}.png"
                 )
+                print('rendering frame', i, frame_img_file, output_file_path)
                 rendered_image = bpy.data.images.load(frame_img_file)
+                print('scaling frame', i, init_image_width, init_image_height)
                 rendered_image.scale(init_image_width, init_image_height)
                 DreamStateOperator.current_frame_idx = i + 1
                 print(
@@ -248,7 +261,6 @@ class GeneratorWorker(Thread):
             scene.frame_set(0)
 
         DreamStateOperator.rendering_from_viewport = False
-        DreamStateOperator.rendering_from_texture = False
         if self.running:
             DreamStateOperator.render_state = RenderState.FINISHED
 
@@ -260,11 +272,11 @@ class DreamRenderOperator(Operator):
     bl_label = "Dream!"
 
     def modal(self, context, event):
-
         settings = context.scene.ds_settings
         output_location = OutputDisplayLocation[settings.output_location]
         ui_context = DreamStateOperator.ui_context
-
+        init_type = get_init_type()
+        prefs = get_preferences()
 
         if DreamStateOperator.render_start_time:
             settings.current_time = time.time() - DreamStateOperator.render_start_time
@@ -274,14 +286,14 @@ class DreamRenderOperator(Operator):
             return {"FINISHED"}
 
         if DreamStateOperator.render_state == RenderState.FINISHED:
+            DreamStateOperator.account = get_account_details(prefs.base_url, prefs.api_key)
+            DreamStateOperator.last_account_check_time = time.time()
             DreamStateOperator.render_state = RenderState.IDLE
             image_tex_area = None
             for area in bpy.context.screen.areas:
                 if area.type == "IMAGE_EDITOR":
                     image_tex_area = area
-            if (
-                output_location == OutputDisplayLocation.TEXTURE_VIEW
-            ):
+            if output_location == OutputDisplayLocation.TEXTURE_VIEW and init_type != InitType.ANIMATION:
                 rendered_image = bpy.data.images.load(
                     DreamStateOperator.last_rendered_image_path
                 )
@@ -326,13 +338,19 @@ class DreamRenderOperator(Operator):
         settings = context.scene.ds_settings
         scene = bpy.context.scene
         ui_context = DreamStateOperator.ui_context
-        init_source = get_init_source()
+        init_type = get_init_type()
         # Ensure there isn't an existing thread with a lock on the render directory.
         DreamStateOperator.kill_render_thread()
-        
-        rendered_dir, generated_images_dir, generated_animation_dir = setup_render_directories()
-        out_dir = generated_animation_dir if init_source == InitSource.EXISTING_VIDEO else generated_images_dir
-        DreamStateOperator.rendered_images_dir = rendered_dir
+        (
+            rendered_dir,
+            generated_images_dir,
+            generated_animation_dir,
+        ) = setup_render_directories()
+        out_dir = (
+            generated_animation_dir
+            if init_type == InitType.ANIMATION
+            else generated_images_dir
+        )
         DreamStateOperator.generated_output_dir = out_dir
         if context.area.type == "IMAGE_EDITOR":
             ui_context = UIContext.IMAGE_EDITOR
@@ -341,30 +359,36 @@ class DreamRenderOperator(Operator):
 
         init_image_width, init_image_height = get_init_image_dimensions(settings, scene)
         render_file_path = scene.render.filepath
-        render_file_type = scene.render.image_settings.file_format
         init_img_paths = []
 
         if DreamStateOperator.rendering_from_viewport:
-            init_source = InitSource.VIEWPORT
-
-        if DreamStateOperator.rendering_from_texture:
-            init_source = InitSource.TEXTURE
+            init_type = InitType.VIEWPORT
 
         # If we are in the image editor, we need to save the image to a temporary file to use for init
-        if init_source == InitSource.TEXTURE:
-            img = context.space_data.image
+        if init_type == InitType.TEXTURE:
+            img = settings.init_texture_ref
             if not img:
-                raise Exception("No image selected!")
-            init_image = copy_image(img)
-            init_image.scale(init_image_width, init_image_height)
-            init_image.save_render(init_img_path)
+                raise Exception("No init texture set")
+            # workaround for render result not having pixels
+            # https://blender.stackexchange.com/questions/2170/how-to-access-render-result-pixels-from-python-script
+            if img.name == "Render Result":
+                img = bpy.data.images['Render Result']
+                rr_path = rendered_dir + "/render_result.png"
+                img.save_render(rr_path, scene=None)
+                init_image = bpy.data.images.load(rr_path)
+                init_image.scale(init_image_width, init_image_height)
+                init_image.save_render(init_img_path)
+            else:
+                init_image = copy_image(img)
+                init_image.scale(init_image_width, init_image_height)
+                init_image.save_render(init_img_path)
             init_img_paths = [init_img_path]
-        
-        if init_source == InitSource.EXISTING_IMAGE:
+
+        if init_type == InitType.TEXTURE:
             init_img_paths = [render_file_path]
 
         # Render 3D view
-        if init_source == InitSource.VIEWPORT:
+        if init_type == InitType.VIEWPORT:
 
             scene.render.filepath = init_img_path
             workspace = bpy.context.workspace.name
@@ -385,20 +409,15 @@ class DreamRenderOperator(Operator):
 
             if res != {"FINISHED"}:
                 raise Exception("Failed to render: {}".format(res))
-        elif init_source == InitSource.EXISTING_VIDEO:
-            render_directory = os.path.dirname(render_file_path)
-            frames_glob = os.path.join(
-                render_directory,
-                "*.{}".format(render_file_type.lower()),
-            )
-            init_img_paths = glob(frames_glob)
+        elif init_type == InitType.ANIMATION:
+            init_img_paths, frame_path = get_anim_images()
         DreamStateOperator.generator_thread = GeneratorWorker(
             scene,
             context,
             ui_context,
             input_img_paths=init_img_paths,
             output_img_directory=out_dir,
-            init_source=init_source,
+            init_type=init_type,
         )
         DreamStateOperator.generator_thread.start()
 
@@ -420,7 +439,6 @@ class DreamStateOperator(Operator):
     total_frame_count = 0
     generator_thread: Thread = None
     rendering_from_viewport = False
-    rendering_from_texture = False
     # Where we put images that are generated after the diffusion step.
     # Either generated_images_dir or generation_animation_dir will be set.
     generated_output_dir = None
@@ -431,6 +449,9 @@ class DreamStateOperator(Operator):
     render_start_time: float = None
 
     sentry_initialized = False
+
+    account: DSAccount = None
+    last_account_check_time = 0
 
     # Cancel any in-progress render and reset the addon state.
     def reset_render_state():
@@ -471,8 +492,24 @@ class DS_OpenOutputFolderOperator(Operator):
             dreamstudio_dir,
             generated_images_dir,
             generated_animation_dir,
-        ) = setup_render_directories()
-        open_folder(dreamstudio_dir)
+        ) = setup_render_directories(clear_anim=False)
+        init_type = get_init_type()
+        if init_type == InitType.ANIMATION:
+            open_folder(generated_animation_dir)
+        else:
+            open_folder(generated_images_dir)
+        return {"FINISHED"}
+
+class DS_UseRenderFolderOperator(Operator):
+    """Use the current Output Path setting in Output Properties as the input folder"""
+    bl_idname = "dreamstudio.set_render_folder"
+    bl_label = "Use Render Folder"
+
+    def execute(self, context):
+        settings = get_settings()
+        scene = context.scene
+        abs_filepath = bpy.path.abspath(scene.render.filepath)
+        settings.init_animation_folder_path = abs_filepath
         return {"FINISHED"}
 
 
@@ -527,3 +564,8 @@ class SelectFileOperator(bpy.types.Operator):
     def invoke(self, context, event):
         context.window_manager.fileselect_add(self)
         return {"RUNNING_MODAL"}
+
+def get_render_texture():
+    for tex in bpy.data.textures:
+        if tex.type == "IMAGE" and tex.name == "Render Result":
+            return tex
